requirements.txt
python-telegram-bot==20.7
beautifulsoup4
requests
matplotlib
apscheduler